feat: update fcitx5 to 5.1.10
Fcitx Input Method Framework v5 (core library dev files)

Issue: https://github.com/deepin-community/sig-deepin-sysdev-team/issues/552
Log: update repo

+++ src/frontend/dbusfrontend/dbusfrontend.cpp
@@ -27,7 +27,12 @@ namespace fcitx {
 
 namespace {
 
-enum { BATCHED_COMMIT_STRING = 0, BATCHED_PREEDIT, BATCHED_FORWARD_KEY };
+enum {
+    BATCHED_COMMIT_STRING = 0,
+    BATCHED_PREEDIT,
+    BATCHED_FORWARD_KEY,
+    BATCHED_DELETE_SURROUNDING
+};
 
 using DBusBlockedEvent = dbus::DBusStruct<uint32_t, dbus::Variant>;
 
@@ -179,15 +184,20 @@ public:
     }
 
     void deleteSurroundingTextImpl(int offset, unsigned int size) override {
-        deleteSurroundingTextDBusTo(name_, offset, size);
+        if (blocked_) {
+            blockedEvents_.emplace_back(
+                BATCHED_DELETE_SURROUNDING,
+                dbus::DBusStruct<int32_t, uint32_t>(offset, size));
+        } else {
+            deleteSurroundingTextDBusTo(name_, offset, size);
+        }
     }
 
     void updateClientSideUIImpl() override {
-        auto preedit =
-            im_->instance()->outputFilter(this, inputPanel().preedit());
-        auto auxUp = im_->instance()->outputFilter(this, inputPanel().auxUp());
-        auto auxDown =
-            im_->instance()->outputFilter(this, inputPanel().auxDown());
+        auto instance = im_->instance();
+        auto preedit = instance->outputFilter(this, inputPanel().preedit());
+        auto auxUp = instance->outputFilter(this, inputPanel().auxUp());
+        auto auxDown = instance->outputFilter(this, inputPanel().auxDown());
         auto candidateList = inputPanel().candidateList();
         int cursorIndex = 0;
 
@@ -209,9 +219,9 @@ public:
                 Text labelText = candidate.hasCustomLabel()
                                      ? candidate.customLabel()
                                      : candidateList->label(i);
-                labelText = im_->instance()->outputFilter(this, labelText);
+                labelText = instance->outputFilter(this, labelText);
                 Text candidateText =
-                    im_->instance()->outputFilter(this, candidate.text());
+                    instance->outputFilter(this, candidate.textWithComment());
                 candidates.emplace_back(std::make_tuple(
                     labelText.toString(), candidateText.toString()));
             }
